+++ vcztools/__init__.py
@@ -1,5 +1,5 @@
 from .bgen import BgenEncoder, write_bgi, write_sample
-from .cli import ViewBgenOptions, ViewPlinkOptions
+from .cli import GroupedCommand, ViewBgenOptions, ViewPlinkOptions
 from .plink import BedEncoder, write_bim, write_fam
 from .provenance import __version__
 from .utils import open_zarr
@@ -7,6 +7,7 @@ from .utils import open_zarr
 __all__ = [
     "BedEncoder",
     "BgenEncoder",
+    "GroupedCommand",
     "ViewBgenOptions",
     "ViewPlinkOptions",
     "__version__",